feat(crm): implement deal access checks and replace useCrmDealAccess hook with utility function
